Art Showcase Page
Added a page for showcasing my artwork, it uses a JSON variable generated by a Python script to automatically populate the page. The only flaw to this right now is that I could not find a way to make my Makefile detect changes in the art folder, so I have to manually run the python script if I modify any art.